+++ package.js
@@ -5,25 +5,25 @@
   Package.describe({
     name: 'xolvio:cucumber',
     summary: 'CucumberJS for Velocity',
-    version: '0.1.0',
+    version: '0.1.1',
     git: 'git@github.com:xolvio/meteor-cucumber.git',
     debugOnly: true
   });
 
   Npm.depends({
-    'cucumber': '0.4.4',
-    'lodash': '2.4.1'
+    'cucumber': '0.4.4'
   });
 
   Package.onUse(function (api) {
 
     api.use([
-      'velocity:core@0.4.2',
-      'velocity:node-soft-mirror@0.0.9',
+      'underscore@1.0.2',
+      'velocity:core@0.4.5',
+      'velocity:node-soft-mirror@0.2.4',
       'velocity:shim@0.0.3'
     ], ['server', 'client']);
     api.use([
-      'velocity:html-reporter@0.3.0-rc.3'
+      'velocity:html-reporter@0.3.2'
     ], 'client');
 
     api.add_files([